refactor(jadestern): components 폴더의 컴포넌트 폴더들을 .js로 변경
