+++ app.js
@@ -1,2 +1,2 @@
-import Header from "./components/Header/app.js";
-import Footer from "./components/Footer/app.js";
+import Header from "./components/Header.js";
+import Footer from "./components/Footer.js";
